requirements.txt
# Install the following directly from the repo before installing all requirements in this file:
#
# git+https://github.com/django-nonrel/django@nonrel-1.4
# git+https://github.com/django-nonrel/djangotoolbox@toolbox-1.4
# git+https://github.com/django-nonrel/mongodb-engine@mongodb-engine-1.4-beta

Django==1.4.5
distribute==0.6.34
django-mongodb-engine==0.4.0
djangotoolbox==0.9.2
pymongo==2.5.2
wsgiref==0.1.2
django-extensions==1.1.1
ipython==0.13.2
six==1.3.0
djangorestframework==2.3.6
